RR-2785 - Refactored "all S&As complete" message integration tests to cover all scenarios for PES
